+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased (wait for DeviceDetectorHelper::getBodyClasses() to be available)
 --------------------
 - Cgh: Minimal HumHub version is now 1.18
 - Chg: SCSS instead of LESS
+- Chg: Remove Custom SCSS field (a migration copies the values to the new HumHub core Appearance form)
 - Chg: Remove the modification of the HumHub SCSS colors variables (we assume all colors are now CSS variables) 
 - Enh: Add body classes from the new `DeviceDetectorHelper` in 1.17
 - Enh: Topbar brand text not centered vertically